+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/viewmodel/channels/ChannelListViewModel.kt
@@ -30,6 +30,7 @@ import io.getstream.chat.android.client.api.models.querysort.QuerySorter
 import io.getstream.chat.android.client.call.toUnitCall
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.models.ChannelMute
+import io.getstream.chat.android.client.models.ConnectionState
 import io.getstream.chat.android.client.models.Filters
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.compose.state.QueryConfig
@@ -39,7 +40,6 @@ import io.getstream.chat.android.compose.state.channels.list.ChannelItemState
 import io.getstream.chat.android.compose.state.channels.list.ChannelsState
 import io.getstream.chat.android.offline.extensions.globalState
 import io.getstream.chat.android.offline.extensions.queryChannelsAsState
-import io.getstream.chat.android.offline.model.connection.ConnectionState
 import io.getstream.chat.android.offline.plugin.state.querychannels.ChannelsStateData
 import io.getstream.chat.android.offline.plugin.state.querychannels.QueryChannelsState
 import io.getstream.logging.StreamLog
@@ -123,12 +123,12 @@ public class ChannelListViewModel(
     /**
      * The state of our network connection - if we're online, connecting or offline.
      */
-    public val connectionState: StateFlow<ConnectionState> = chatClient.globalState.connectionState
+    public val connectionState: StateFlow<ConnectionState> = chatClient.clientState.connectionState
 
     /**
      * The state of the currently logged in user.
      */
-    public val user: StateFlow<User?> = chatClient.globalState.user
+    public val user: StateFlow<User?> = chatClient.clientState.user
 
     /**
      * Gives us the information about the list of channels mutes by the current user.
@@ -139,7 +139,7 @@ public class ChannelListViewModel(
      * Builds the default channel filter, which represents "messaging" channels that the current user is a part of.
      */
     private fun buildDefaultFilter(): Flow<FilterObject> {
-        return chatClient.globalState.user.map(Filters::defaultChannelListFilter).filterNotNull()
+        return chatClient.clientState.user.map(Filters::defaultChannelListFilter).filterNotNull()
     }
 
     /**
@@ -313,7 +313,7 @@ public class ChannelListViewModel(
     public fun loadMore() {
         logger.d { "Loading more channels" }
 
-        if (chatClient.globalState.isOffline()) return
+        if (chatClient.clientState.isOffline) return
         val currentConfig = QueryConfig(
             filters = filterFlow.value ?: return,
             querySort = querySortFlow.value