+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/offline/plugin/logic/internal/LogicRegistry.kt
@@ -21,8 +21,8 @@ import io.getstream.chat.android.client.api.models.FilterObject
 import io.getstream.chat.android.client.api.models.QueryChannelsRequest
 import io.getstream.chat.android.client.api.models.querysort.QuerySorter
 import io.getstream.chat.android.client.extensions.cidToTypeAndId
-import io.getstream.chat.android.client.logger.ChatLogger
 import io.getstream.chat.android.client.models.Channel
+import io.getstream.chat.android.client.setup.state.ClientState
 import io.getstream.chat.android.offline.plugin.logic.channel.internal.ChannelLogic
 import io.getstream.chat.android.offline.plugin.logic.channel.internal.ChannelStateLogic
 import io.getstream.chat.android.offline.plugin.logic.channel.internal.ChannelStateLogicImpl
@@ -37,6 +37,7 @@ import io.getstream.chat.android.offline.plugin.state.global.internal.MutableGlo
 import io.getstream.chat.android.offline.plugin.state.global.internal.toMutableState
 import io.getstream.chat.android.offline.plugin.state.querychannels.internal.toMutableState
 import io.getstream.chat.android.offline.repository.builder.internal.RepositoryFacade
+import io.getstream.logging.StreamLog
 import kotlinx.coroutines.runBlocking
 import java.util.concurrent.ConcurrentHashMap
 
@@ -49,6 +50,7 @@ import java.util.concurrent.ConcurrentHashMap
 internal class LogicRegistry internal constructor(
     private val stateRegistry: StateRegistry,
     private val globalState: MutableGlobalState,
+    private val clientState: ClientState,
     private val userPresence: Boolean,
     private val repos: RepositoryFacade,
     private val client: ChatClient,
@@ -80,7 +82,7 @@ internal class LogicRegistry internal constructor(
     fun channel(channelType: String, channelId: String): ChannelLogic {
         return channels.getOrPut(channelType to channelId) {
             val mutableState = stateRegistry.channel(channelType, channelId).toMutableState()
-            val stateLogic = ChannelStateLogicImpl(mutableState, globalState, SearchLogic(mutableState))
+            val stateLogic = ChannelStateLogicImpl(mutableState, globalState, clientState, SearchLogic(mutableState))
 
             ChannelLogic(
                 repos = repos,
@@ -149,7 +151,7 @@ internal class LogicRegistry internal constructor(
     internal companion object {
         private var instance: LogicRegistry? = null
 
-        private val logger = ChatLogger.get("LogicRegistry")
+        private val logger = StreamLog.getLogger("Chat:LogicRegistry")
 
         /**
          * Creates and returns new instance of LogicRegistry.
@@ -164,22 +166,32 @@ internal class LogicRegistry internal constructor(
          *
          * @throws IllegalStateException if instance is not null.
          */
+        @Suppress("LongParameterList")
         internal fun create(
             stateRegistry: StateRegistry,
             globalState: MutableGlobalState,
+            clientState: ClientState,
             userPresence: Boolean,
             repos: RepositoryFacade,
             client: ChatClient,
         ): LogicRegistry {
             if (instance != null) {
-                logger.logE(
+                logger.e {
                     "LogicRegistry instance is already created. " +
                         "Avoid creating multiple instances to prevent ambiguous state. Use LogicRegistry.get()"
-                )
+                }
             }
-            return LogicRegistry(stateRegistry, globalState, userPresence, repos, client).also { logicRegistry ->
-                instance = logicRegistry
-            }
+            return LogicRegistry(
+                stateRegistry,
+                globalState,
+                clientState,
+                userPresence,
+                repos,
+                client
+            )
+                .also { logicRegistry ->
+                    instance = logicRegistry
+                }
         }
 
         /**