+++ stream-chat-android-ui-components-sample/src/main/kotlin/io/getstream/chat/ui/sample/feature/channel/add/group/select_name/AddGroupChannelSelectNameViewModel.kt
@@ -22,7 +22,6 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.models.User
-import io.getstream.chat.android.offline.extensions.globalState
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import java.util.UUID
@@ -45,7 +44,7 @@ class AddGroupChannelSelectNameViewModel : ViewModel() {
         _state.value = State.Loading
         viewModelScope.launch(Dispatchers.Main) {
             val currentUserId =
-                ChatClient.instance().globalState.user.value?.id ?: error("User must be set before create new channel!")
+                ChatClient.instance().clientState.user.value?.id ?: error("User must be set before create new channel!")
             val result = ChatClient.instance()
                 .createChannel(
                     channelType = CHANNEL_TYPE_MESSAGING,